+++ src/app/components/post/postContent.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { PropTypes } from 'prop-types';
 import { connect } from 'react-redux';
-import { withRouter } from 'react-router'
+import { withRouter } from 'react-router';
 import { fetchPost } from './actions';
 
 class PostContent extends Component {