+++ src/app/components/post/postList.jsx
@@ -12,20 +12,25 @@ class PostList extends Component {
   render() {
     return (
       <div>
-        {this.props.posts.map(post => (
-          <PostItem {...post} key={post.postId} />
-        ))}
+        { // FIXME: 这里这么写非常非常不优雅
+          // TODO: tag和author现在是只有id的，需要把具体内容从对应entities拿出来
+          this.props.posts.result.map((postId) => {
+          const post = this.props.posts.entities.posts[postId];
+          return <PostItem {...post} key={post.postId} />;
+        })}
       </div>
     );
   }
 }
 
 PostList.propTypes = {
-  posts: PropTypes.array.isRequired,
+  posts: PropTypes.object.isRequired,
   onPostListLoad: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = state => ({
+  // TODO: get all posts list from state比如 getOnlyPostInfo(state)
+  // TODO: and extract tag and author info from entities with id
   posts: state.posts,
 });